requirements.txt
streamlit>=1.40,<2
pandas>=2.2,<3
openpyxl>=3.1,<4
python-dateutil>=2.9,<3
python-dotenv>=1.0,<2
